+++ src/components/TextClassification/index.js
@@ -1,12 +1,12 @@
-import React, { useState, useEffect, useRef } from "react"
+import React, { useState } from "react"
 import getTaskDescription from "../../utils/get-task-description.js"
 import SampleContainer from "../SampleContainer"
 import NLPAnnotator from "react-nlp-annotate/components/NLPAnnotator"
-import * as datasetHelper from "../../utils/dataset-helper"
-
+import useSampleText from "../../utils/file-handlers/use-sample-text"
 export const TextClassification = (props) => {
   const [currentSampleIndex, changeCurrentSampleIndex] = useState(0)
   const { annotation } = props.samples[currentSampleIndex]
+  const { textToShow } = useSampleText(props, currentSampleIndex)
   // TODO remove legacy support for availableLabels
   if (!props.interface.labels && props.interface.availableLabels) {
     props.interface.labels = props.interface.availableLabels
@@ -24,19 +24,6 @@ export const TextClassification = (props) => {
       typeof l === "string" ? { id: l, description: l, displayName: l } : l
     )
     .map((l) => (!l.displayName ? { ...l, displayName: l.id } : l))
-
-  const [textToShow, changeTextToShow] = useState("")
-  const oldText = useRef()
-  useEffect(() => {
-    if (oldText.current !== textToShow) {
-      datasetHelper
-        .getTextfromSample(props.samples[currentSampleIndex])
-        .then((result) => {
-          changeTextToShow(result)
-        })
-      oldText.current = textToShow
-    }
-  }, [props.samples, currentSampleIndex, textToShow])
 
   return (
     <SampleContainer
