+++ src/utils/dataset-helper/erase-annotation-one-sample.js
@@ -1,0 +1,8 @@
+import isEmpty from "lodash/isEmpty"
+export default (sample) => {
+  let Newsample = sample
+  if (!isEmpty(Newsample.annotation)) {
+    delete Newsample["annotation"]
+  }
+  return Newsample
+}
