+++ src/utils/dataset-helper/get-sample-url.js
@@ -1,0 +1,10 @@
+export default (sample) => {
+  return (
+    sample.imageUrl ||
+    sample.videoUrl ||
+    sample.audioUrl ||
+    sample.pdfUrl ||
+    sample.textUrl ||
+    undefined
+  )
+}
