+++ src/utils/file-handlers/use-aws-cognito/index.js
@@ -2,10 +2,9 @@
 
 import { useEffect, useRef, useCallback } from "react"
 import UpdateAWSStorage from "./update-aws-storage"
-import * as datasetHelper from "../../dataset-helper"
 import isEmpty from "lodash/isEmpty"
 import useAuth from "../../auth-handlers/use-auth"
-
+import getFilesDifferences from "../../dataset-helper/get-files-differences"
 const workingInterfaces = [
   "video_segmentation",
   "image_classification",
@@ -24,7 +23,7 @@ export default ({ file }) => {
   const shouldUpdateAWSStorage = useCallback(() => {
     if (!isLoggedIn || authProvider !== "cognito") return
 
-    var changes = datasetHelper.fileHasChanged(lastObjectRef.current, file)
+    var changes = getFilesDifferences(lastObjectRef.current, file)
     if (
       isEmpty(file) ||
       (!changes.content.samples && !changes.fileName) ||
